+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+* server: added support for SSE (Server-Side Events) to `HttpExchange`
 * json: improve parsing of complex types with parameters
 * jdbc: introduce `NullableId` for entities with null ids until they are persisted.
 * jdbc: introduce `UpdatableEntity` for optimistic locking in `BaseRepository.save`